# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIPRIMA API Documentation",
        "description": "API Documentation untuk Sistem Informasi Prima (SIPRIMA)",
        "contact": {
            "email": "support@siprima.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.siprima.digitaltech.my.id",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/lokasis": {
            "get": {
                "tags": [
                    "Lokasi"
                ],
                "summary": "Get all lokasis",
                "description": "Mendapatkan daftar semua lokasi",
                "operationId": "71c88bd4dd958593035a0f7c0ada0257",
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan data lokasis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "Gedung A"
                                                    },
                                                    "alamat": {
                                                        "type": "string",
                                                        "example": "Jl. Sudirman No. 123"
                                                    },
                                                    "latitude": {
                                                        "type": "string",
                                                        "example": "-6.200000"
                                                    },
                                                    "longitude": {
                                                        "type": "string",
                                                        "example": "106.816666"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lokasi"
                ],
                "summary": "Create new lokasi",
                "description": "Membuat lokasi baru",
                "operationId": "f0da28d483398e019e2d2f14398703b5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "alamat"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Gedung A"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Sudirman No. 123"
                                    },
                                    "latitude": {
                                        "type": "string",
                                        "example": "-6.200000"
                                    },
                                    "longitude": {
                                        "type": "string",
                                        "example": "106.816666"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lokasi berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lokasi berhasil dibuat"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dengan format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}